Lean more on inherited packages. Set PYTHONPATH.
Removed a lot of module installations for things that are already
in the base image.  I originally wanted to make sure we had
updated versions of important modules (for example, pandas), but
this turned out to be unnecessary for most tasks.  So now we only
install modules that aren't already present.

Also added "/home/jovyan/work/lib" to the default PYTHONPATH
in the image.  If /home/jovyan/work is a mounted volume, it
allows the user an easy way to drop modules into a directory
and have them available in all notebooks.