Fixed import warning with requests, and switched to 'pip install --upgrade'
The 'requests' module was throwing warnings about incompatible versions for chardet and urllib3, so we explicitly
had conda install that in order to be sure we had the most recent version.

Also, when installing modules via pip, the --upgrade flag will now make sure they are updated to the latest version,
even if older versions already exist on the system.